requirements.txt
streamlit
python-dotenv
google-genai
numpy
pandas
plotly
openpyxl
databricks-sql-connector